+++ zholobov.oleg/common/tree/node.hpp
